requirements.txt
streamlit
gspread
google-generativeai
Pillow
